AppiumUtilClassModifiedVersion
In this class i have added a new way of starting and stopping the appium server (using AppiumServiceBuilder class) which will work across all platforms that Appium supports. 

Also, if you are using Eclipse and facing this error "cannot run program adb error=2 no such file or directory", when running the standalone adb commands through Eclipse, i have provided a solution through this code. It's not particularly an impressive solution, but a work around or now, until the time i figure out how to make Eclipse drive the ADB available in the my machine. 

When trying to clone and/ or download this class and use, please update the package name, its currently not in sync with this project structure. Please feel free to update it.